Created authentication middleware
